Overwriting rotational with corotational
