--- src/BPM/bond_bpm_corotational.h
@@ -1,61 +1,0 @@
-/* -*- c++ -*- ----------------------------------------------------------
-   LAMMPS - Large-scale Atomic/Molecular Massively Parallel Simulator
-   https://www.lammps.org/, Sandia National Laboratories
-   LAMMPS development team: developers@lammps.org
-
-   Copyright (2003) Sandia Corporation.  Under the terms of Contract
-   DE-AC04-94AL85000 with Sandia Corporation, the U.S. Government retains
-   certain rights in this software.  This software is distributed under
-   the GNU General Public License.
-
-   See the README file in the top-level LAMMPS directory.
-------------------------------------------------------------------------- */
-
-#ifdef BOND_CLASS
-// clang-format off
-BondStyle(bpm/corotational,BondBPMCorotational);
-// clang-format on
-#else
-
-#ifndef LMP_BOND_BPM_COROTATIONAL_H
-#define LMP_BOND_BPM_COROTATIONAL_H
-
-#include "bond_bpm.h"
-
-namespace LAMMPS_NS {
-
-class BondBPMCorotational : public BondBPM {
- public:
-  BondBPMCorotational(class LAMMPS *);
-  ~BondBPMCorotational() override;
-  void compute(int, int) override;
-  void coeff(int, char **) override;
-  void init_style() override;
-  void settings(int, char **) override;
-  void write_restart(FILE *) override;
-  void read_restart(FILE *) override;
-  void write_restart_settings(FILE *) override;
-  void read_restart_settings(FILE *) override;
-  double single(int, double, int, int, double &) override;
-
- protected:
-  double *Kr, *Ks, *Kt, *Kb, *gr, *gs, *gt, *gb;
-  double *Fcr, *Fcs, *Tct, *Tcb;
-  int smooth_flag, normalize_flag;
-  int frame_style, damping_style;
-
-  double corotational_forces(int, int, int, double *, double *, double *, double *,
-                                   double *, double &, double *);
-  double standard_forces(int, int, int, double *, double *, double *, double *,
-                                   double *, double &, double *);
-  void dem_damping_forces(int, int, int, double *, double *, double *, double *);
-
-  void allocate();
-  void store_data();
-  double store_bond(int, int, int);
-};
-
-}    // namespace LAMMPS_NS
-
-#endif
-#endif
